# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.aiLearningTermRequest:
    properties:
      category:
        type: string
      enabled:
        type: boolean
      groupId:
        type: string
      language:
        type: string
      relevanceLevel:
        type: string
      signalRole:
        type: string
      term:
        type: string
      topicKey:
        type: string
      weight:
        type: number
    type: object
...
  main.aiReassessmentJobView:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      failedCount:
        type: integer
      id:
        type: string
      processedCount:
        type: integer
      skippedCount:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      totalCount:
        type: integer
    type: object
...
  main.connectorAccountRequest:
    properties:
      label:
        type: string
      platform:
        type: string
    type: object
  main.connectorAccountView:
    properties:
      externalAccountId:
        type: string
      id:
        type: string
      label:
        type: string
      lastConnectedAt:
        type: string
      lastError:
        type: string
      lastSyncAt:
        type: string
      platform:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  main.connectorQRView:
    properties:
      accountId:
        type: string
      connected:
        type: boolean
      expiresAt:
        type: string
      lastError:
        type: string
      leaseKind:
        type: string
      platform:
        type: string
      qr:
        type: string
      queueLength:
        type: integer
      queuePosition:
        type: integer
      requestId:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      waitReason:
        type: string
    type: object
...
  main.knowledgeRebuildJobView:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      failedCount:
        type: integer
      id:
        type: string
      processedCount:
        type: integer
      replaceExisting:
        type: boolean
      skippedCount:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      totalCount:
        type: integer
      updatedAt:
        type: string
    type: object
...
  main.replayJobView:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      failedCount:
        type: integer
      from:
        type: string
      groupIds:
        items:
          type: string
        type: array
      id:
        type: string
      includeMedia:
        type: boolean
      processedCount:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      to:
        type: string
      totalCount:
        type: integer
    type: object
  main.replayRequest:
    properties:
      from:
        type: string
      groupIds:
        items:
          type: string
        type: array
      includeMedia:
        type: boolean
      to:
        type: string
    type: object
...
  main.threadReassessmentJobView:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      failedCount:
        type: integer
      id:
        type: string
      processedCount:
        type: integer
      skippedCount:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      totalCount:
        type: integer
    type: object
info:
  contact:
    name: CONXTOR project
    url: https://github.com/conxtor/groupmaster
  description: |-
    REST API for ingesting, analyzing, and presenting selected WhatsApp and Telegram group conversations.
    The public deployment serves this API below the /api reverse-proxy prefix.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/conxtor/groupmaster
  title: CONXTOR Messaging Group Intelligence API
  version: 0.1.0
paths:
...
  /admin/ai-learning/{id}:
    delete:
      parameters:
      - description: Learning term UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete AI learning term
      tags:
      - ai-learning
    patch:
      consumes:
      - application/json
      parameters:
      - description: Learning term UUID
        in: path
        name: id
        required: true
        type: string
      - description: Learning term
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.aiLearningTermRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update AI learning term
      tags:
      - ai-learning
...
  /admin/ai-learning/reassessment:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.aiReassessmentJobView'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List message reassessment jobs
      tags:
      - ai-learning
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.aiReassessmentJobView'
            type: array
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.aiReassessmentJobView'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Reassess all messages
      tags:
      - ai-learning
...
  /admin/ai-learning/thread-reassessment:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.threadReassessmentJobView'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List thread reassessment jobs
      tags:
      - ai-learning
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.threadReassessmentJobView'
            type: array
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.threadReassessmentJobView'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Reassess inferred conversation threads
      tags:
      - ai-learning
...
  /admin/knowledge/topics/rebuild:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.knowledgeRebuildJobView'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List Knowledge Base rebuild jobs
      tags:
      - administration
    post:
      consumes:
      - application/json
      parameters:
      - description: Rebuild options
        in: body
        name: body
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.knowledgeRebuildJobView'
            type: array
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.knowledgeRebuildJobView'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Rebuild Knowledge Base
      tags:
      - administration
...
  /connectors/accounts:
    get:
      parameters:
      - description: Administrators may include all accounts
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.connectorAccountView'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List connector accounts
      tags:
      - connectors
    post:
      consumes:
      - application/json
      parameters:
      - description: Connector account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.connectorAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.connectorAccountView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create connector account
      tags:
      - connectors
...
  /connectors/accounts/{accountId}/qr:
    get:
      parameters:
      - description: Connector account UUID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.connectorQRView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get or start connector QR onboarding
      tags:
      - connectors
    post:
      consumes:
      - application/json
      parameters:
      - description: Connector account UUID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.connectorQRView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Start connector QR onboarding
      tags:
      - connectors
...
  /replays:
    get:
      consumes:
      - application/json
      parameters:
      - description: Administrators may include all users' jobs
        in: query
        name: all
        type: boolean
      - description: Replay range and groups
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.replayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.replayJobView'
            type: array
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.replayJobView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List or create replay jobs
      tags:
      - replays
    post:
      consumes:
      - application/json
      parameters:
      - description: Replay range and groups
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.replayRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.replayJobView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create replay job
      tags:
      - replays
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: Authenticated CONXTOR session cookie.
    in: cookie
    name: wagi_session
    type: apiKey
swagger: "2.0"